--- p01/src/main/java/javva/p01/Main.java
@@ -1,7 +1,0 @@
-package javva.p01;
-
-public final class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
