+++ p01/src/main/java/practice/p01/Main.java
@@ -1,0 +1,7 @@
+package practice.p01;
+
+public final class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
