--- p01/src/test/java/javva/p01/simple/JavaKoKo.java
@@ -1,17 +1,0 @@
-package javva.p01.simple;
-
-import java.text.MessageFormat;
-import java.util.Random;
-import java.util.StringTokenizer;
-import java.util.concurrent.ThreadLocalRandom;
-
-public final class JavaKoKo {
-
-    public static void main(String[] args) {
-        StringTokenizer tokenizer = new StringTokenizer("", ";!, ");
-        while (tokenizer.hasMoreElements()) {
-            tokenizer.nextToken();
-        }
-
-    }
-}
