+++ p01/src/test/java/ru/mirea/practice/simple/JavaKoKo.java
@@ -1,0 +1,14 @@
+package ru.mirea.practice.simple;
+
+import java.util.StringTokenizer;
+
+public final class JavaKoKo {
+
+    public static void main(String[] args) {
+        StringTokenizer tokenizer = new StringTokenizer("", ";!, ");
+        while (tokenizer.hasMoreElements()) {
+            tokenizer.nextToken();
+        }
+
+    }
+}
