--- students/0000001/0000001-p01/src/main/java/javva/s0000001/Main.java
@@ -1,7 +1,0 @@
-package javva.s0000001;
-
-public final class Main {
-    public static void main(String[] args) {
-        System.out.println("первая практическая работа!");
-    }
-}
