+++ students/0000001/0000001-p01/src/main/java/ru/mirea/practice/s0000001/Main.java
@@ -1,0 +1,7 @@
+package ru.mirea.practice.s0000001;
+
+public final class Main {
+    public static void main(String[] args) {
+        System.out.println("первая практическая работа!");
+    }
+}
