--- students/21K0709/21K0709-p01/src/main/java/javva/s21k0709/ArgClass.java
@@ -1,10 +1,0 @@
-package javva.s21k0709;
-
-//Третья задача 1 практической работы
-public class ArgClass {
-    public static void main(String[] args) {
-        for (String str : args) {
-            System.out.println("аргумент = " + str);
-        }
-    }
-}
