--- students/21K0709/21K0709-p02/src/main/java/javva/task10/s21k0709/HowMany.java
@@ -1,19 +1,0 @@
-package javva.task10.s21k0709;
-
-import java.util.Scanner;
-
-public class HowMany {
-    public static void main(String[] args) {
-        Scanner sc = new Scanner(System.in);
-        System.out.print("Введите строку слов: ");
-        int col = 1;
-        String name;
-        name = sc.nextLine();
-        for (int i = 0; i < name.length(); i++) {
-            if (name.charAt(i) == ' ') {
-                col++;
-            }
-        }
-        System.out.print("Кол-во слов: " + col);
-    }
-}
