--- students/21K0709/21K0709-p02/src/main/java/javva/task2/s21k0709/Ball.java
@@ -1,44 +1,0 @@
-package javva.task2.s21k0709;
-
-public class Ball {
-    private double x = 0.0;
-    private double y = 0.0;
-
-    public Ball(double x, double y) {
-        this.x = x;
-        this.y = y;
-    }
-
-    public Ball() {
-    }
-
-    public double getX() {
-        return x;
-    }
-
-    public void setX(double x) {
-        this.x = x;
-    }
-
-    public double getY() {
-        return y;
-    }
-
-    public void setY(double y) {
-        this.y = y;
-    }
-
-    public void setXY(double x, double y) {
-        this.x = x;
-        this.y = y;
-    }
-
-    public void move(double x, double y) {
-
-    }
-
-    @Override
-    public String toString() {
-        return "Координаты: " + "x = " + x + " y = " + y;
-    }
-}
