--- students/21K0709/21K0709-p02/src/main/java/javva/task3/s21k0709/Circle.java
@@ -1,36 +1,0 @@
-package javva.task3.s21k0709;
-
-public class Circle {
-    private double r;
-    private String color;
-    private Point p = new Point();
-
-    public Circle() {
-    }
-
-    public Circle(Point p, double r) {
-        this.p = p;
-        this.r = r;
-    }
-
-    public double getR() {
-        return r;
-    }
-
-    public void setR(double r) {
-        this.r = r;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public void setColor(String color) {
-        this.color = color;
-    }
-
-    @Override
-    public String toString() {
-        return "Координаты: " + "p:  " + p + " r = " + r;
-    }
-}
