--- students/21K0709/21K0709-p02/src/main/java/javva/task3/s21k0709/Tester.java
@@ -1,30 +1,0 @@
-package javva.task3.s21k0709;
-
-import java.util.Scanner;
-
-public class Tester {
-    public static void main(String[] args) {
-        Scanner sc = new Scanner(System.in);
-        int j = 0; // кол-во эл массива
-        do {
-            System.out.println("Введите кол-во элементов массива: ");
-            if (sc.hasNextInt()) {
-                j = sc.nextInt(); //считывает целое число из поток ввода и сохраняет в переменную
-            }
-        } while (j < 0);
-        Circle[] mas = new Circle[j];
-        for (int i = 0; i < mas.length; i++) {
-            System.out.println("Введите x: ");
-            float x = sc.nextFloat();
-            System.out.println("Введите y: ");
-            float y = sc.nextFloat();
-            System.out.println("Введите r: ");
-            float r = sc.nextFloat();
-            mas[i] = new Circle(new Point(x, y), r);
-        }
-        for (int i = 0; i < mas.length; i++) {
-            System.out.println(mas[i]);
-        }
-    }
-
-}
