+++ students/21K0709/21K0709-p02/src/main/java/ru/mirea/practice/task3/s21k0709/Point.java
@@ -1,0 +1,38 @@
+package ru.mirea.practice.task3.s21k0709;
+
+public class Point {
+    private float x = 0;
+    private float y = 0;
+
+
+    public Point() {
+    }
+
+    ;
+
+    public Point(float x, float y) {
+        this.x = x;
+        this.y = y;
+    }
+
+    public float getX() {
+        return x;
+    }
+
+    public void setX(float x) {
+        this.x = x;
+    }
+
+    public float getY() {
+        return y;
+    }
+
+    public void setY(float y) {
+        this.y = y;
+    }
+
+    @Override
+    public String toString() {
+        return "x = " + x + " y = " + y;
+    }
+}
